Moved login and registration responsibility to the Login and Registration classes, respectively. Created an AccountAccessAttempt class that holds the generic information that both the Login and Registration classes have. Login and Registration classes are now subclasses of AccountAccessAttempt.
